change the fb into UpdateCommend
