+++ app/src/main/res/layout/activity_qr.xml
@@ -74,41 +74,7 @@
             app:layout_constraintVertical_bias="0.502"
             app:srcCompat="@drawable/border_line" />
 
-        <Button
-            android:id="@+id/button1"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="36dp"
-            android:background="@drawable/save_button"
-            android:text="Button"
-            app:layout_constraintBottom_toTopOf="@+id/imageView2"
-            app:layout_constraintStart_toEndOf="@+id/button2"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintVertical_bias="0.666" />
 
-        <Button
-            android:id="@+id/button2"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginStart="32dp"
-            android:background="@drawable/save_button"
-            android:text="Button"
-            app:layout_constraintBottom_toTopOf="@+id/imageView2"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintVertical_bias="0.666" />
-
-        <Button
-            android:id="@+id/button4"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_marginEnd="36dp"
-            android:background="@drawable/save_button"
-            android:text="Button"
-            app:layout_constraintBottom_toTopOf="@+id/imageView2"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintVertical_bias="0.666" />
 
 
     </androidx.constraintlayout.widget.ConstraintLayout>
@@ -120,7 +86,7 @@
         android:layout_gravity="center_horizontal"
         android:layout_weight="8"
         android:dividerHeight="10.0sp"
-        android:divider= "#F6E587"
+        android:divider= "#FEFFD3"
         android:listSelector="#87d5e2"
         android:orientation="vertical"
         android:layout_marginHorizontal="20dp"
